Optimize terrain heightmap generation
The first observation is that chunk.HeightMap generates heightmaps that
are one level higher than what we do. This means the rough heightmap is
off by one, and by lowering the y level for the block search by one, we
shave off one block test for every (x, z) coordinate pair in the entire
box.

The second observation is that some blocks are way more prevalent than
others, when testing blocks for being a terrain block. Checking block
IDs in order, from most prevalent to least prevalent, means
isTerrainBlock() can return earlier the more frequent the block.

For my first test world, these are the most frequent block IDs for the
blocks tested with isTerrainBlock(), when generating the height map:

    88k water (Block ID 9)
    58k air (id 0)
    43k grass block (id 2)
    30k + 14k leaves (ids 18 and 161)
    2.4k sand (id 12)
    1.7k clay (id 82)
    1.4k grass plant (id 31)

Other block IDs hardly appeared at all. This will of course vary
depending on what biomes are present, but it gives an indication of what
blocks to expect. Tests should be carried out for larger test sets, and
other worlds, in order to determine what order to check block IDs for
good overall performance of terrain heightmap generation.

The third observation is that some of the terrain block IDs tested
against virutally never appear. This includes ores that can only spawn
below a certain y coordinate, blocks that rarely spawns and only in rare
biomes, or blocks that simply never generate as part of any terrain in
any biome.

The fourth observation is that some ranges of block IDs can be tested in
bulk, instead of individually one and one. E.g. for the cost of checking
just sand and gravel one can additionally get three types of ore, and
stone, grass block, dirt and cobblestone combined is only one single
comparison when the only lower ID (air, 0) is already tested for.