Simplify and improve travel distance calculation
Most important is the change from a two dimensional binary array for
marking visited nodes, to using a set. This shaves off a significant
portion of the runtime. The early exit introduced in the last commit
lost its effect and was removed. Also removed other unnecessary code
from the same function.

Note that removal of ``if thusFar >= timeout:'', lines 357-358, means
``if newEstimate <= timeout:'' on line 379 is now necessary for the
timeout condition to work. Timeout relies on the ``edge'' heap queue to
exhaust, and has done so for the past commits. Earlier it depended on a
"too large estimate" to pop up from the queue.