Improve canal finding algorithm
For the main part of the algorithm, finding candidate crossings for
building a canal, change from iterating over a two dimensional array
to using a priority queue. Runtime reduced to a tenth.

Some work still can be done, especially figuring out what is happening
with the edge of the map. (Apparently there are some edge cases there,
pun not intended.) The code could need some cleanup, for instance some
values / state information is unnecessary with the new approach.

Bottleneck now is travelDistance(), which is responsible for about 80 %
of findCanalLocations() running time, or 40 % of total execution time.